[function] 数据与 view 打通

+++ android/app/src/main/kotlin/com/example/flutter_dynamic_weather/map/MinuteViewModel.kt
@@ -88,7 +88,7 @@ class MinuteViewModel : ViewModel() {
 
                 weatherDesc = WeatherUtils.getWeatherDesc(realtime.result?.realtime?.skycon) ?: ""
                 temp = realtime.result?.realtime?.temperature?.toInt()?.toString()
-                aqiDesc = WeatherUtils.getAqiDesc(realtime?.result?.realtime?.airQuality?.aqi?.chn)
+                aqiDesc = WeatherUtils.getAqiDesc(realtime?.result?.realtime?.airQuality?.aqi?.chn.toInt())
                         ?: ""
             }
             if (minuteData != null) {
